requirements.txt
aiortc>=1.9.0
av>=11.0.0
numpy>=2.0.0
requests>=2.32.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
pillow>=10.3.0
python-dotenv>=1.0.1
comfyui-frontend-package>=1.32.5

